+++ changelog.md
@@ -1,4 +1,4 @@
-# Dec 10th '23 V0.0.5
+# Dec 10th '23 V0.0.9
 
     - init
     - Added main menu cards
@@ -7,3 +7,7 @@
     - Added current day store (wip)
     - Fixed day store
     - Cleaned out some bloat and added back button to first page
+    - Added DayPage component for the day's pages (wip)
+    - Removed some redundant code (selectedDay prop on card and daypage)
+    - Changed "description" in days data to "subtitle" and added "description" which is the long text description given on the page
+    - Simplified the first day page to be an else statement and added selectedDayIndex to dynamically grab the days data to pass to the daypage component